requirements.txt
fastapi==0.100.0
hypercorn==0.14.4
google-api-python-client==2.26.1
python-dotenv==0.19.1
requests
